Import japanese word wrap dynamically
